House Service Test

+++ src/main/java/com/mahallem/dto/Response/HouseResponse.java
@@ -18,7 +18,7 @@ public class HouseResponse {
 
     private String name;
 
-    private long grade;
+    private Double grade;
 
     private GeoLocation geoLocation;
